Tratamento de excessões

+++ src/main/java/com/services/CRUD_test/services/DriverServices.java
@@ -5,10 +5,15 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.services.CRUD_test.entities.Driver;
 import com.services.CRUD_test.repository.DriverRepository;
+import com.services.CRUD_test.services.exceptions.DataBaseException;
+import com.services.CRUD_test.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class DriverServices implements Serializable{
@@ -22,21 +27,37 @@ public class DriverServices implements Serializable{
 
     public Driver findById(Long Id){
         Optional <Driver> dr = repository.findById(Id);
-        return dr.orElseThrow(() -> new RuntimeException()); //*TODO SUBSTITUIÇÃO DE RUNTIMEEXCEPTION POR EXCESSÃO PEROSNALIZADA */
+        return dr.orElseThrow(() -> new ResourceNotFoundException(Id)); 
     }
 
     public Driver insert(Driver dr){
         return repository.save(dr);
     }
 
-    public void deleteById(Long Id){  //*TODO CRIAR EXCESSÕES */
-        repository.deleteById(Id); 
+    public void deleteById(Long Id){  
+        try{
+            if (!repository.existsById(Id)) {
+                throw new ResourceNotFoundException(Id);
+            }
+            repository.deleteById(Id);
+        }
+        catch (ResourceNotFoundException e){
+            throw new ResourceNotFoundException(Id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DataBaseException(e.getMessage());
+        }
+    
     }
 
     public Driver update(Long Id, Driver obj){
-        Driver dr = repository.getReferenceById(Id);  //*TODO CRIAR EXCESSÕES */
+        try{
+        Driver dr = repository.getReferenceById(Id); 
         updateDriver(dr, obj);
         return repository.save(dr);
+        }catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException(Id);
+        }
         
     }
